Move brightness slider next to power controls
